+++ src/semantic.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
-#include "schema.hpp"
-#include "storage.hpp"
-
 #include <memory>
 #include <tekla/db1/process.hpp>
 #include <tekla/db1/result.hpp>
+
+#include "schema.hpp"
+#include "storage.hpp"
 
 namespace tekla::db1::detail {
 
@@ -13,6 +13,9 @@ namespace tekla::db1::detail {
     std::shared_ptr<const ModelStorage> storage, const Schema& schema,
     const ProcessRequest& request);
 [[nodiscard]] Result<ProcessStream> make_relation_stream(
+    std::shared_ptr<const ModelStorage> storage, const Schema& schema,
+    const ProcessRequest& request);
+[[nodiscard]] Result<ProcessStream> make_semantic_relation_stream(
     std::shared_ptr<const ModelStorage> storage, const Schema& schema,
     const ProcessRequest& request);